# packages/api/schema/openapi.yaml
openapi: 3.0.3
info:
  title: Campus & Clubs API
  version: 1.0.0
  description: Campus & Clubs Backend API
paths:
...
  /api/v1/accounts/auth/all/:
    get:
      operationId: list_users
      description: Returns a paginated list of users.
      summary: List users
      parameters:
      - in: query
        name: ordering
        schema:
          type: string
        description: Sort field, e.g. `-date_joined`, `username`.
      - in: query
        name: page
        schema:
          type: integer
        description: Page number.
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: search
        schema:
          type: string
        description: Search by username, first name, or last name.
      tags:
      - Users
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserProfileList'
          description: Paginated list of users.
        '401':
          description: Unauthorized.
    post:
      operationId: list_users_2
      description: Returns a paginated list of users.
      summary: List users
      parameters:
      - in: query
        name: ordering
        schema:
          type: string
        description: Sort field, e.g. `-date_joined`, `username`.
      - in: query
        name: page
        schema:
          type: integer
        description: Page number.
      - in: query
        name: search
        schema:
          type: string
        description: Search by username, first name, or last name.
      tags:
      - Users
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/UserProfileRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UserProfileRequest'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserProfileList'
          description: Paginated list of users.
        '401':
          description: Unauthorized.
...
components:
  schemas:
...
    BlankEnum:
      enum:
      - ''
...
    GenderEnum:
      enum:
      - male
      - female
      - other
      type: string
      description: |-
        * `male` - Male
        * `female` - Female
        * `other` - Other
    NullEnum:
      enum:
      - null
    PaginatedUserProfileList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/UserProfile'
...
    PreferredEmailEnum:
      enum:
      - email
      - professional_email
      type: string
      description: |-
        * `email` - Personal Email
        * `professional_email` - Professional Email
...
    StatusEnum:
      enum:
      - online
      - away
      - dnd
      type: string
      description: |-
        * `online` - Online
        * `away` - Away
        * `dnd` - Do Not Disturb
...
    TypeEnum:
      enum:
      - student
      - faculty
      - staff
      - alumni
      - other
      type: string
      description: |-
        * `student` - Student
        * `faculty` - Faculty
        * `staff` - Staff
        * `alumni` - Alumni
        * `other` - Other
...
    UserProfile:
      type: object
      description: Detailed user profile with club, post, and follow information
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          readOnly: true
          title: Email address
        professional_email:
          type: string
          format: email
          readOnly: true
          nullable: true
        institute:
          type: string
          readOnly: true
        institute_id:
          type: string
          readOnly: true
        url:
          type: string
          readOnly: true
        gender:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/GenderEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        student_id:
          type: string
          nullable: true
        year:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        level:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        type:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/TypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        preferred_email:
          allOf:
          - $ref: '#/components/schemas/PreferredEmailEnum'
          description: |-
            Which email address to use for notifications

            * `email` - Personal Email
            * `professional_email` - Professional Email
        profile_picture_url:
          type: string
          readOnly: true
        avatar:
          type: string
          readOnly: true
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
          maxLength: 255
        website:
          nullable: true
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        date_of_birth:
          type: string
          format: date
          nullable: true
        email_verified:
          type: boolean
          readOnly: true
        is_private:
          type: boolean
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          description: |-
            User's current online status

            * `online` - Online
            * `away` - Away
            * `dnd` - Do Not Disturb
        is_status_manual:
          type: boolean
          description: Whether the status was set manually by the user
        club_count:
          type: integer
          description: Get club count from property
          readOnly: true
        clubs:
          type: string
          readOnly: true
        clubs_url:
          type: string
          readOnly: true
        user_post_count:
          type: integer
          readOnly: true
        total_posts_count:
          type: integer
          readOnly: true
        posts_url:
          type: string
          readOnly: true
        follower_count:
          type: integer
          readOnly: true
        following_count:
          type: integer
          readOnly: true
        pending_requests_count:
          type: integer
          readOnly: true
        followers_url:
          type: string
          readOnly: true
        following_url:
          type: string
          readOnly: true
        is_following:
          type: boolean
          nullable: true
          description: Is current user following this user?
          readOnly: true
        is_followed_by:
          type: boolean
          nullable: true
          description: Is this user following current user?
          readOnly: true
        is_mutual:
          type: boolean
          nullable: true
          description: Are they mutual followers?
          readOnly: true
        follow_status:
          nullable: true
          description: Get follow status (pending, accepted, None)
          readOnly: true
        can_view_profile:
          type: boolean
          description: Can current user view this profile?
          readOnly: true
        likes_given:
          type: integer
          readOnly: true
        comments_made:
          type: integer
          readOnly: true
        shares_made:
          type: integer
          readOnly: true
        likes_received:
          type: integer
          readOnly: true
        last_active:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        last_login:
          type: string
          format: date-time
          readOnly: true
          nullable: true
      required:
      - avatar
      - can_view_profile
      - club_count
      - clubs
      - clubs_url
      - comments_made
      - created_at
      - email
      - email_verified
      - follow_status
      - follower_count
      - followers_url
      - following_count
      - following_url
      - id
      - institute
      - institute_id
      - is_followed_by
      - is_following
      - is_mutual
      - last_login
      - likes_given
      - likes_received
      - pending_requests_count
      - posts_url
      - professional_email
      - profile_picture_url
      - shares_made
      - total_posts_count
      - updated_at
      - url
      - user_post_count
      - username
    UserProfileRequest:
      type: object
      description: Detailed user profile with club, post, and follow information
      properties:
        username:
          type: string
          minLength: 1
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
        last_name:
          type: string
          maxLength: 150
        gender:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/GenderEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        student_id:
          type: string
          nullable: true
        year:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        level:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        type:
          nullable: true
          oneOf:
          - $ref: '#/components/schemas/TypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        preferred_email:
          allOf:
          - $ref: '#/components/schemas/PreferredEmailEnum'
          description: |-
            Which email address to use for notifications

            * `email` - Personal Email
            * `professional_email` - Professional Email
        bio:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
          maxLength: 255
        website:
          nullable: true
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        date_of_birth:
          type: string
          format: date
          nullable: true
        is_private:
          type: boolean
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          description: |-
            User's current online status

            * `online` - Online
            * `away` - Away
            * `dnd` - Do Not Disturb
        is_status_manual:
          type: boolean
          description: Whether the status was set manually by the user
        last_active:
          type: string
          format: date-time
          nullable: true
      required:
      - username
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT